Split 'protocol' into an entire crate
This module previously was compiled into both 'sand' and the main
bandsocks lib, by using an include!() to copy the text. This is a bit
nicer, making a shared crate for the common functionality. The crate
itself still needs to be copied/symlinked during the build, since it
must exist simultaneously in two different workspaces.